# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Solunex Lab Backend",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/portal/login":  {
                                        "post":  {
                                                     "tags":  [
                                                                  "portal"
                                                              ],
                                                     "summary":  "Portal Login",
                                                     "operationId":  "portal_login_portal_login_post",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/PortalLogin"
                                                                                                                             }
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Successful Response",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/PortalTokenOut"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                                       "422":  {
                                                                                   "description":  "Validation Error",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/portal/results":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "portal"
                                                               ],
                                                      "summary":  "Portal List Results",
                                                      "operationId":  "portal_list_results_portal_results_get",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "authorization",
                                                                             "in":  "header",
                                                                             "required":  false,
                                                                             "schema":  {
                                                                                            "anyOf":  [
                                                                                                          {
                                                                                                              "type":  "string"
                                                                                                          },
                                                                                                          {
                                                                                                              "type":  "null"
                                                                                                          }
                                                                                                      ],
                                                                                            "title":  "Authorization"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Successful Response",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "type":  "array",
                                                                                                                                            "items":  {
                                                                                                                                                          "$ref":  "#/components/schemas/PortalResultItem"
                                                                                                                                                      },
                                                                                                                                            "title":  "Response Portal List Results Portal Results Get"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                },
                                                                        "422":  {
                                                                                    "description":  "Validation Error",
                                                                                    "content":  {
                                                                                                    "application/json":  {
                                                                                                                             "schema":  {
                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                        }
                                                                                                                         }
                                                                                                }
                                                                                }
                                                                    }
                                                  }
                                      },
                  "/portal/results/{result_id}/pdf":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "portal"
                                                                               ],
                                                                      "summary":  "Portal Download Pdf",
                                                                      "operationId":  "portal_download_pdf_portal_results__result_id__pdf_get",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "result_id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "integer",
                                                                                                            "title":  "Result Id"
                                                                                                        }
                                                                                         },
                                                                                         {
                                                                                             "name":  "authorization",
                                                                                             "in":  "header",
                                                                                             "required":  false,
                                                                                             "schema":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "null"
                                                                                                                          }
                                                                                                                      ],
                                                                                                            "title":  "Authorization"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "Successful Response",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {

                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                },
                                                                                        "422":  {
                                                                                                    "description":  "Validation Error",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "PortalLogin":  {
                                                           "properties":  {
                                                                              "phone":  {
                                                                                            "type":  "string",
                                                                                            "title":  "Phone"
                                                                                        },
                                                                              "patient_no":  {
                                                                                                 "type":  "string",
                                                                                                 "title":  "Patient No"
                                                                                             }
                                                                          },
                                                           "type":  "object",
                                                           "required":  [
                                                                            "phone",
                                                                            "patient_no"
                                                                        ],
                                                           "title":  "PortalLogin"
                                                       },
                                       "PortalResultItem":  {
                                                                "properties":  {
                                                                                   "id":  {
                                                                                              "type":  "integer",
                                                                                              "title":  "Id"
                                                                                          },
                                                                                   "test_type_id":  {
                                                                                                        "type":  "integer",
                                                                                                        "title":  "Test Type Id"
                                                                                                    },
                                                                                   "template_id":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "Template Id"
                                                                                                   },
                                                                                   "status":  {
                                                                                                  "type":  "string",
                                                                                                  "title":  "Status"
                                                                                              },
                                                                                   "created_at":  {
                                                                                                      "type":  "string",
                                                                                                      "format":  "date-time",
                                                                                                      "title":  "Created At"
                                                                                                  }
                                                                               },
                                                                "type":  "object",
                                                                "required":  [
                                                                                 "id",
                                                                                 "test_type_id",
                                                                                 "template_id",
                                                                                 "status",
                                                                                 "created_at"
                                                                             ],
                                                                "title":  "PortalResultItem"
                                                            },
                                       "PortalTokenOut":  {
                                                              "properties":  {
                                                                                 "token":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Token"
                                                                                           },
                                                                                 "expires_at":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "date-time",
                                                                                                    "title":  "Expires At"
                                                                                                }
                                                                             },
                                                              "type":  "object",
                                                              "required":  [
                                                                               "token",
                                                                               "expires_at"
                                                                           ],
                                                              "title":  "PortalTokenOut"
                                                          },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}